El tamany màxim dels logs ara s'especifica amb un long long
De pas, s'ha augmentat fins a 5GB.

+++ src/lib/util/debug.h
@@ -32,7 +32,7 @@ void debug_done(void);
  * to the 'debug' function. If a file name is given, an implicit call to
  * 'debug_assign_file' is performed. If max_size is not 0, the file will be
  * truncated if the limit is rebased. */
-int debug_new_category(char *filename, unsigned int max_size);
+int debug_new_category(char *filename, long long max_size);
 
 /* Switch the status of a debugging category. By default, the
  * debugging messages are on, while there is an opened file to be dumped into. */
